Revert "settings react -> nextjs"
