+++ frontend_nextjs/package.json
@@ -12,10 +12,8 @@
     "axios": "^0.26.1",
     "bootstrap": "^5.1.3",
     "js-cookie": "^3.0.1",
-    "next": "^12.1.0",
+    "next": "12.1.0",
     "next-auth": "^4.3.1",
-    "next.js": "^1.0.3",
-    "nextjs": "^0.0.3",
     "react": "17.0.2",
     "react-bootstrap": "^2.2.1",
     "react-dom": "17.0.2",
