--- frontend_nextjs/styles/manageusers.css
@@ -1,11 +1,0 @@
-
-.dropdown-button{
-    min-width:134px;
-    min-height:38px;
-}
-
-.table-manage-users{
-    overflow-y:auto;
-    width: 100%;
-    max-height: 60%;
-}
